recommendations/requirements.txt
grpcio-tools==1.43.0
grpc-interceptor==0.13.1
pytest==7.0.0